% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{aGreatFunction}
\alias{aGreatFunction}
\title{Add together two numbers}
\usage{
aGreatFunction(x, y)
}
\description{
This is a description
}
\details{
@param x A number.
@param y A number.
@return The sum of \code{x} and \code{y}
@export
@examples
aGreatFunction{1,1}
AGreatFunction{10,1}
}
